--- src/components/About.js
@@ -1,9 +1,0 @@
-/* eslint-disable no-unused-vars */
-import React from 'react';
-import { Row, Col, Form, Button } from 'react-bootstrap';
-
-function About() {
-  return <section>About</section>;
-}
-
-export default About;
