--- src/components/NewsArticles.js
@@ -1,6 +1,0 @@
-import React from 'react';
-
-// eslint-disable-next-line react/function-component-definition
-const NewsArticles = () => <div>News Articles</div>;
-
-export default NewsArticles;
