+++ src/components/tests/AddJournal.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import AddJournal from './index';
+import AddJournal from '../index';
 
 test('Add Journal renders the form correctly', () => {
   const { getByText, getByLabelText } = render(<AddJournal />);
