+++ src/components/tests/SignUp.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import SignUp from '../pages/SignUp';
+import SignUp from '../index';
 
 test('renders the form correctly', () => {
   const { getByText, getByLabelText } = render(<SignUp />);
