+++ src/pages/Journal/Journal.css
