+++ src/pages/Manifesto/Manifesto.css
