+++ src/pages/Manifesto/Manifesto.js
@@ -2,14 +2,14 @@
 /* eslint-disable react/function-component-definition */
 /* eslint-disable import/order */
 import React from 'react';
-import '../styles/Manifesto.css';
+import './Manifesto.css';
 import { Container, Row, Col, Image } from 'react-bootstrap';
-import StandingImg from '../assets/human-2.webp';
-import SittingImg from '../assets/human-1.webp';
-import ImgOne from '../assets/manifesto-img-1.webp';
-import WalkingImg from '../assets/walking-man.webp';
-import PeersImg from '../assets/peers.webp';
-import HandsIcon from '../assets/hands-icon.webp';
+import StandingImg from '../../assets/human-2.webp';
+import SittingImg from '../../assets/human-1.webp';
+import ImgOne from '../../assets/manifesto-img-1.webp';
+import WalkingImg from '../../assets/walking-man.webp';
+import PeersImg from '../../assets/peers.webp';
+import HandsIcon from '../../assets/hands-icon.webp';
 import { ImTwitter, ImMail, ImGithub } from 'react-icons/im';
 
 const Manifesto = () => (
