+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export { default as Contact } from './Contact/Contact';
+export { default as Journal } from './Journal/Journal';
+export { default as Manifesto } from './Manifesto/Manifesto';
+export { default as Home } from './Home';
